+++ hy/cmdline.py
@@ -4,6 +4,9 @@
 
 from __future__ import print_function
 
+import colorama
+colorama.init()
+
 import argparse
 import code
 import ast